Move the stop-gap interpreter to a sub-module.
This allows us to reduce the number of always-public exports, so that we can use
the "testing" feature more easily at the level of structs.
